third_party/WebKit/LayoutTests/fast/forms/checkbox/checkbox-checked-state-affected-by-default-state.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
test (function() {
    var el = document.createElement("input");
    el.setAttribute("type", "checkbox");
    el.defaultChecked = true;
    el.checked = true;
    el.defaultChecked = false;
    assert_true(el.checked);
}, "This test check that checked state of checkbox once explicitly set is not affected by default state.");
</script>
</body>
</html>
